+++ src/components/member/Join2.js
@@ -1,477 +1,338 @@
+import { Modal } from "@mui/material";
 import axios from "axios";
 import { useEffect, useRef, useState } from "react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
 import Swal from "sweetalert2";
+import ChevronRightIcon from "@mui/icons-material/ChevronRight";
 
-const Join2 = () => {
+const Join3 = () => {
   const navigate = useNavigate(); // 페이지 이동을 위한 navigate 훅
   const location = useLocation(); // 현재 위치에서 전달된 데이터를 가져오는 훅
+  const [isFormValid, setIsFormValid] = useState(false); // 폼 유효성 상태
   const { email, code } = location.state || {}; // 이메일과 코드 데이터를 가져옴 (없으면 기본값으로 빈 객체 설정)
+  const [isOpen, setIsOpen] = useState(false);
 
   console.log(email);
   console.log(code);
 
-  useEffect(() => {
-    if (!code) {
-      Swal.fire({
-        title: "코드 누락",
-        text: "인증 코드가 누락되었습니다. 메인 페이지로 이동합니다.",
-        icon: "error",
-      }).then(() => {
-        navigate("/"); // 메인 페이지로 이동
-      });
-    }
-  }, [code, navigate]);
+  const [check1, setCheck1] = useState(false); // '약관 전체 동의' 상태
+  const [check2, setCheck2] = useState(false);
+  const [check3, setCheck3] = useState(false);
+  const [check4, setCheck4] = useState(false);
+  const [check5, setCheck5] = useState(false);
+  const [check6, setCheck6] = useState(false);
+  const [check7, setCheck7] = useState(false);
 
-  // 회원 정보를 관리하는 상태 변수
-  const [member, setMember] = useState({
-    memberEmail: "", // 기본 이메일 설정 (이메일 인증 페이지에서 전달된 이메일로 업데이트됨)
-    memberPw: "", // 비밀번호 상태
-    memberNickname: "", // 닉네임 상태
-    memberPhone: "", // 휴대폰 번호 상태
-    memberBirth: "", // 생년월일 상태
-    memberGender: "", // 성별 상태
-  });
-
-  // useEffect를 사용해 email 값이 변경되면 member 상태 업데이트
-  useEffect(() => {
-    if (email) {
-      setMember((prevState) => ({ ...prevState, memberEmail: email }));
-    }
-  }, [email]);
-
-  // 비밀번호 확인 입력 값 상태 관리
-  const [memberPwRe, setMemberPwRe] = useState("");
-  const [isFormValid, setIsFormValid] = useState(false); // 폼 유효성 상태
-
-  // 비밀번호 확인 입력값 업데이트 핸들러
-  const inputMemberPwRe = (e) => {
-    setMemberPwRe(e.target.value);
+  // '약관 전체 동의' 체크박스 클릭 시
+  const handleAllCheckChange = (e) => {
+    const checked = e.target.checked;
+    setCheck1(checked);
+    setCheck2(checked);
+    setCheck3(checked);
+    setCheck4(checked);
+    setCheck5(checked);
+    setCheck6(checked);
+    setCheck7(checked);
   };
 
-  // 회원 정보 입력 필드 업데이트 핸들러
-  const inputMemberData = (e) => {
-    const name = e.target.name; // 입력 필드의 name 속성
-    const inputData = e.target.value; // 입력 필드의 값
-    setMember({ ...member, [name]: inputData }); // 해당 필드의 값 상태 업데이트
-  };
-
-  // 비밀번호 일치 여부를 체크하고, 메시지 표시를 위한 ref
-  const pwMsgRef = useRef(null);
-
-  // 비밀번호 일치 여부를 확인하는 함수
-  const checkPw = () => {
-    pwMsgRef.current.classList.remove("input-ok"); // 이전 상태에서 input-ok, input-error, input-error2 클래스 제거
-    pwMsgRef.current.classList.remove("input-error");
-
-    if (member.memberPw && memberPwRe) {
-      // 비밀번호와 비밀번호 확인 값이 null 또는 빈 문자열이 아니면
-      if (member.memberPw === memberPwRe) {
-        // 비밀번호가 일치하면 "input-ok" 클래스 추가하고 일치 메시지 표시
-        pwMsgRef.current.classList.add("input-ok");
-        pwMsgRef.current.innerText = "비밀번호가 일치합니다.";
-      } else {
-        // 비밀번호가 일치하지 않으면 "input-error" 클래스 추가하고 불일치 메시지 표시
-        pwMsgRef.current.classList.add("input-error");
-        pwMsgRef.current.innerText = "비밀번호가 일치하지 않습니다.";
-      }
+  // 개별 체크박스 클릭 시 '약관 전체 동의' 상태를 체크
+  const handleIndividualChange = () => {
+    if (check2 && check3 && check4 && check5 && check6 && check7) {
+      setCheck1(true);
     } else {
-      // 비밀번호 또는 비밀번호 확인이 입력되지 않은 경우
-      pwMsgRef.current.classList.remove("input-ok");
-      pwMsgRef.current.classList.add("input-error");
-      pwMsgRef.current.innerText = "비밀번호를 입력해주세요.";
+      setCheck1(false);
     }
   };
 
-  // 닉네임 중복 체크 및 유효성 검사
-  const [nicknameCheck, setNicknameCheck] = useState(0);
-  const checkNickname = () => {
-    const idReg = /^[a-zA-Z0-9가-힣]+$/; // 닉네임에 사용할 수 있는 문자: 한글, 영어 대소문자, 숫자
-    if (idReg.test(member.memberNickname)) {
-      // 정규식에 맞으면 중복 체크 요청
-      axios
-        .get(
-          `${process.env.REACT_APP_BACK_SERVER}/member/exists?memberNickname=${member.memberNickname}`
-        )
-        .then((res) => {
-          console.log(res);
-          if (res.data === 0) {
-            setNicknameCheck(1); // 사용 가능한 닉네임
-          } else {
-            setNicknameCheck(3); // 이미 사용중인 닉네임
-          }
-        })
-        .catch((err) => {
-          console.log(err); // 에러 발생 시 콘솔 출력
-        });
+  const handleToggle = () => {
+    setIsOpen((prev) => !prev); // 클릭 시 토글
+  };
+
+  // 폼 유효성 체크: 필수 항목들이 모두 체크되었는지 확인
+  useEffect(() => {
+    if (check2 && check3 && check4) {
+      setIsFormValid(true);
     } else {
-      setNicknameCheck(2); // 유효하지 않은 닉네임
+      setIsFormValid(false);
     }
-  };
-
-  // 휴대폰 번호 유효성 검사 및 형식 변경
-  const [phoneError, setPhoneError] = useState(0);
-  const checkPhone = () => {
-    const phoneRegex = /^(010-\d{4}-\d{4})$/; // 010-0000-0000 형식의 정규표현식
-    const formattedPhone = member.memberPhone.replace(/[^0-9]/g, ""); // 숫자만 추출
-    let formattedPhoneWithDash = "";
-    if (formattedPhone.length === 11) {
-      formattedPhoneWithDash = `${formattedPhone.slice(
-        0,
-        3
-      )}-${formattedPhone.slice(3, 7)}-${formattedPhone.slice(7)}`;
-    } else {
-      formattedPhoneWithDash = member.memberPhone;
-    }
-
-    setMember({ ...member, memberPhone: formattedPhoneWithDash });
-
-    if (!phoneRegex.test(formattedPhoneWithDash)) {
-      setPhoneError(2); // 잘못된 형식일 경우 에러
-    } else {
-      setPhoneError(1); // 유효한 형식일 경우 에러 메시지 제거
-    }
-  };
-
-  // 회원가입 요청 함수
-  const joinMember = () => {
-    if (!member.memberBirth) {
-      Swal.fire({
-        title: "생년월일 선택 오류",
-        text: "생년월일을 선택해주세요.",
-        icon: "error",
-        confirmButtonColor: "var(--main2)",
-      });
-      return;
-    }
-
-    console.log(member); // 회원 정보 확인
-    axios
-      .post(`${process.env.REACT_APP_BACK_SERVER}/member/join`, member) // 서버에 회원가입 요청
-      .then((res) => {
-        Swal.fire({
-          title: "회원가입 성공!",
-          text: "회원가입이 완료되었습니다.",
-          icon: "success",
-          confirmButtonColor: "var(--main2)",
-        }).then(() => {
-          navigate("/"); // 회원가입 성공 후 홈으로 리디렉션
-        });
-      })
-      .catch((err) => {
-        Swal.fire({
-          title: "오류",
-          text: "회원가입 중 오류가 발생했습니다. 다시 시도해주세요.",
-          icon: "error",
-          confirmButtonColor: "var(--main2)",
-        });
-      });
-  };
-
-  // 생년월일 관련 상태 변수 및 생성
-  const [year, setYear] = useState("");
-  const [month, setMonth] = useState("");
-  const [day, setDay] = useState("");
-  const currentYear = new Date().getFullYear(); // 현재 연도
-  const years = Array.from(
-    { length: currentYear - 1900 + 1 },
-    (_, i) => currentYear - i
-  ); // 연도 목록
-  const months = Array.from({ length: 12 }, (_, i) => i + 1); // 월 목록
-  const getDaysInMonth = (year, month) => {
-    const daysInMonth = new Date(year, month, 0).getDate(); // 해당 월의 일수 구하기
-    return Array.from({ length: daysInMonth }, (_, i) => i + 1); // 일수 목록
-  };
-  const [days, setDays] = useState([]);
-  useEffect(() => {
-    if (year && month) {
-      const updatedDays = getDaysInMonth(year, month);
-      setDays(updatedDays); // 월에 맞는 일 목록 업데이트
-      if (!updatedDays.includes(parseInt(day))) {
-        setDay(updatedDays[0] || "");
-      }
-    }
-  }, [year, month, day]);
-
-  // 생년월일 값이 설정되면 member 상태에 반영
-  useEffect(() => {
-    if (year !== "" && month !== "" && day !== "") {
-      const formattedMonth = String(month).padStart(2, "0");
-      const formattedDay = String(day).padStart(2, "0");
-      const birthString = `${year}${formattedMonth}${formattedDay}`; // 생년월일 문자열 형식
-
-      setMember((prevState) => ({ ...prevState, memberBirth: birthString }));
-    }
-  }, [year, month, day]);
-
-  // member 상태 업데이트 후 바로 실행되는 effect
-  useEffect(() => {
-    if (member.memberBirth) {
-      // memberBirth가 변경될 때마다 콘솔 로그로 출력
-      console.log(member);
-    }
-  }, [member.memberBirth]); // memberBirth 상태에 의존
-
-  useEffect(() => {
-    const isValid =
-      member.memberPw &&
-      memberPwRe &&
-      member.memberPw === memberPwRe &&
-      member.memberNickname &&
-      nicknameCheck === 1 &&
-      member.memberPhone &&
-      phoneError === 1 &&
-      member.memberBirth &&
-      member.memberGender;
-
-    setIsFormValid(isValid); // 폼 유효성 상태 업데이트
-  }, [member, memberPwRe, nicknameCheck, phoneError, year, month, day]);
+  }, [check2, check3, check4]);
 
   return (
     <section className="section">
-      {code ? (
-        <>
-          <div className="join-title">
-            <h2>필수 정보 입력</h2>
-            <p>가입을 위해 필수 정보를 입력해주세요.</p>
+      {/*    {code ? (*/}
+      <>
+        <div className="join-title">
+          <h2>이용약관</h2>
+        </div>
+
+        <div className="join-agree-wrap">
+          <div className="join-agree-first">
+            <input
+              type="checkbox"
+              id="check1"
+              checked={check1}
+              onChange={handleAllCheckChange}
+            />
+            <label htmlFor="check1">
+              약관 전체동의 <p>선택항목 포함</p>
+            </label>
           </div>
 
-          <div className="join-wrap">
-            <form
-              onSubmit={(e) => {
-                e.preventDefault(); // 폼 제출 시 기본 동작 방지
-                joinMember(); // 회원가입 함수 호출
+          <nav className="join-agree-table">
+            <ul>
+              <li>
+                <input
+                  type="checkbox"
+                  id="check2"
+                  checked={check2}
+                  onChange={(e) => {
+                    setCheck2(e.target.checked);
+                    handleIndividualChange();
+                  }}
+                />
+                <label htmlFor="check2">(필수) 이용약관</label>
+                <ChevronRightIcon
+                  className="join-chevron-icon"
+                  onClick={handleToggle}
+                  style={{
+                    transition: "transform 0.3s ease",
+                    transform: isOpen ? "rotate(90deg)" : "rotate(0deg)",
+                    marginLeft: "auto",
+                    cursor: "pointer",
+                  }}
+                />
+                <div
+                  className={`agree-details ${isOpen ? "open" : ""}`}
+                  style={{
+                    maxHeight: isOpen ? "200px" : "0", // 펼쳐지거나 접힐 때의 높이
+                    transform: isOpen ? "translateY(0)" : "translateY(-20px)", // 아래로 떨어지는 효과
+                    transition:
+                      "max-height 0.5s ease, transform 0.3s ease, opacity 0.5s ease", // 부드러운 애니메이션
+                    overflow: "hidden", // 내용이 넘치지 않도록 숨김 처리
+                  }}
+                >
+                  <div className="agree-content">
+                    <div
+                      className="agree-content2"
+                      style={{ marginTop: "0px", marginLeft: "15px" }}
+                    >
+                      <h3>제 1조 (목적)</h3>
+                      <p>
+                        이 약관은 주식회사 여기어때컴퍼니(이하 "회사"라 함)가
+                        운영하는 종합 여행·여가 플랫폼(이하 "플랫폼"이라
+                        한다)에서 제공하는 예약 관련 서비스(이하 "서비스"라
+                        한다)를 이용함에 있어 "회사"와 "이용자"의 권리․의무 및
+                        책임사항을 규정함을 목적으로 합니다.
+                      </p>
+                      <h3>제 2조 (정의)</h3>
+                      <p>
+                        ① "플랫폼"이란 "이용자"가 컴퓨터 등 정보통신설비를
+                        이용하여 "서비스"를 이용할 수 있도록 "회사"가 제공하는
+                        가상의 영업장을 말하며, 아울러 "플랫폼"을 운영하는
+                        사업자의 의미로도 사용합니다.
+                      </p>
+                      <p>
+                        ② "이용자"란 "플랫폼"을 통하여 이 약관에 따라 제공하는
+                        서비스를 받는 회원 및 비회원을 말합니다.
+                      </p>
+                      <p>
+                        ③ "회원"이라 함은 "플랫폼"에 회원등록을 한 자로서,
+                        계속적으로 "플랫폼"이 제공하는 서비스를 이용할 수 있는
+                        자를 말합니다.
+                      </p>
+                      <p>
+                        ④ "비회원"이라 함은 회원에 가입하지 않고 "플랫폼"이
+                        제공하는 서비스를 이용하는 자를 말합니다.
+                      </p>
+                      <p>
+                        ⑤ 서비스수수료란 "이용자"가 "홈&빌라" 서비스를 이용할 때
+                        "플랫폼"에 지급하는 수수료를 말합니다.
+                      </p>
+                      <p>
+                        ⑥ "게시물"이란 "이용자"가 "서비스" 이용 시 게시하는
+                        리뷰,사진, 글, 댓글, 동영상, 프로필 사진 등의 콘텐츠
+                        일체를 의미합니다.
+                      </p>
+                      <h3>제 3조 (약관 등의 명시와 설명 및 개정)</h3>
+                      <p>
+                        ① "회사"는 이 약관의 내용과 상호 및 대표자 성명, 영업소
+                        소재지 주소(소비자의 불만을 처리할 수 있는 곳의 주소를
+                        포함), 전화번호, E-mail주소, 사업자등록번호, 통신판매업
+                        신고번호, 개인정보보호책임자등을 "이용자"가 쉽게 알 수
+                        있도록 "플랫폼"의 초기 서비스화면(전면)에 게시합니다.
+                        다만, 약관의 내용은 "이용자"가 연결화면을 통하여 볼 수
+                        있도록 할 수 있습니다.
+                      </p>
+                      <p>
+                        ② "회사"는 「전자상거래 등에서의 소비자보호에 관한
+                        법률」, 「약관의 규제에 관한 법률」, 「전자문서 및
+                        전자거래기본법」, 「전자금융거래법」, 「전자서명법」,
+                        「정보통신망 이용촉진 및 정보보호 등에 관한 법률」,
+                        「소비자기본법」 등 관련 법을 위배하지 않는 범위에서 이
+                        약관을 개정할 수 있습니다.
+                      </p>
+                      <p>
+                        ③ "회사"가 약관을 개정할 경우에는 적용일자 및 개정사유를
+                        명시하여 현행약관과 함께 플랫폼의 초기화면에 그 적용일자
+                        7일 이전부터 적용일자 전일까지 공지합니다. 다만,
+                        "이용자"에게 불리하게 약관내용을 변경하는 경우에는
+                        최소한 30일 이상의 사전 유예기간을 두고 공지합니다. 이
+                        경우 "회사"는 개정 전 내용과 개정 후 내용을 "이용자"가
+                        알기 쉽도록 표시합니다.
+                      </p>
+                      <p>
+                        ④ "회원"은 변경된 약관에 동의하지 않을 권리가 있으며,
+                        동의하지 않을 경우에는 서비스이용을 중단하고 탈퇴할 수
+                        있습니다.
+                      </p>
+                      <p>
+                        ⑤ "회원"이 전④항에 따라 약관에 대한 반대의사를 표시하지
+                        않고 "서비스"를 이용한 경우에는 약관에 동의한 것으로
+                        봅니다.
+                      </p>
+                      <p>
+                        ⑥ 이 약관에서 정하지 아니한 사항과 이 약관의 해석에
+                        관하여는 전자상거래 등에서의 소비자보호에 관한 법률,
+                        약관의 규제 등에 관한 법률 및 관계법령 또는 상관례에
+                        따릅니다.
+                      </p>
+                      <h3>제 4조 (회원가입)</h3>
+                      <p>
+                        ① "이용자"는 "플랫폼"이 정한 절차에 따라 이 약관에
+                        동의한다는 의사표시를 함으로서 회원가입을 신청합니다.
+                      </p>
+                      <p>
+                        ② "회사"는 제①항과 같이 회원으로 가입할 것을 신청한
+                        "이용자" 중 다음 각 호에 해당하지 않는 한 회원으로
+                        등록합니다. 1. 회원자격 상실 후 24시간이 경과하지 않은
+                        경우 2. 등록 내용에 타인의 정보를 사용한 경우 3. 만 14세
+                        미만의 아동이 신청하는 경우
+                      </p>
+                      <p>
+                        ③ 회원가입계약의 성립 시기는 "회사"의 승낙이 회원에게
+                        도달한 시점으로 합니다.
+                      </p>
+                      <p>
+                        ④ 회원은 회원가입 시 등록한 사항에 변경이 있는 경우,
+                        상당한 기간 이내에 "플랫폼"에 대하여 회원정보 수정하거나
+                        E-mail 등의 방법으로 그 변경사항을 알려야 합니다.
+                      </p>
+                      <p>
+                        ⑤ "회사"는 관련법령에 따라 필요한 경우 별도의 성인인증
+                        절차를 실시할 수 있습니다.
+                      </p>
+                    </div>
+                  </div>
+                </div>
+              </li>
+
+              <li>
+                <input
+                  type="checkbox"
+                  id="check3"
+                  checked={check3}
+                  onChange={(e) => {
+                    setCheck3(e.target.checked);
+                    handleIndividualChange();
+                  }}
+                />
+                <label htmlFor="check3">(필수) 만 14세 이상 확인</label>
+                <ChevronRightIcon className="join-chevron-icon" />
+              </li>
+              <li>
+                <input
+                  type="checkbox"
+                  id="check4"
+                  checked={check4}
+                  onChange={(e) => {
+                    setCheck4(e.target.checked);
+                    handleIndividualChange();
+                  }}
+                />
+                <label htmlFor="check4">
+                  (필수) 개인정보 수집 및 이용 동의
+                </label>
+                <ChevronRightIcon className="join-chevron-icon" />
+              </li>
+              <li>
+                <input
+                  type="checkbox"
+                  id="check5"
+                  checked={check5}
+                  onChange={(e) => {
+                    setCheck5(e.target.checked);
+                    handleIndividualChange();
+                  }}
+                />
+                <label htmlFor="check5">
+                  (선택) 개인정보 수집 및 이용 동의
+                </label>
+                <ChevronRightIcon className="join-chevron-icon" />
+              </li>
+              <li>
+                <input
+                  type="checkbox"
+                  id="check6"
+                  checked={check6}
+                  onChange={(e) => {
+                    setCheck6(e.target.checked);
+                    handleIndividualChange();
+                  }}
+                />
+                <label htmlFor="check6">(선택) 마케팅 알림 수신 동의</label>
+                <ChevronRightIcon className="join-chevron-icon" />
+              </li>
+              <li>
+                <input
+                  type="checkbox"
+                  id="check7"
+                  checked={check7}
+                  onChange={(e) => {
+                    setCheck7(e.target.checked);
+                    handleIndividualChange();
+                  }}
+                />
+                <label htmlFor="check7">
+                  (선택) 위치기반 서비스 이용약관 동의
+                </label>
+                <ChevronRightIcon className="join-chevron-icon" />
+              </li>
+            </ul>
+          </nav>
+          <div className="join-button-box2">
+            <button
+              type="submit"
+              className={`join-button ${isFormValid ? "" : "disabled"}`}
+              disabled={!isFormValid}
+              style={{
+                pointerEvents: !isFormValid ? "none" : "auto", // 버튼 비활성화 시 클릭 불가
+                backgroundColor: isFormValid ? "#30c272" : "white", // 비활성화 시 배경색 흰색으로
+                color: isFormValid ? "white" : "#d3d3d3", // 비활성화 시 글자색 여린 회색으로
+              }}
+              onMouseEnter={(e) => {
+                if (isFormValid) {
+                  e.target.style.backgroundColor = "#166139"; // hover 시 배경색 변경
+                }
+              }}
+              onMouseLeave={(e) => {
+                if (isFormValid) {
+                  e.target.style.backgroundColor = "#30c272"; // hover 끝난 후 원래 배경색으로 복구
+                }
               }}
             >
-              {/* 비밀번호 입력 */}
-              <div className="input-wrap">
-                <div className="input-title">
-                  <label htmlFor="memberPw">비밀번호</label>
-                </div>
-                <div className="input-item">
-                  <input
-                    type="password"
-                    name="memberPw"
-                    id="memberPw"
-                    value={member.memberPw}
-                    onChange={inputMemberData}
-                    onBlur={checkPw} // 입력이 끝나면 비밀번호 일치 여부 확인
-                    placeholder="최소 8자 이상"
-                  />
-                </div>
-              </div>
-
-              {/* 비밀번호 확인 입력 */}
-              <div className="input-wrap">
-                <div className="input-title">
-                  <label htmlFor="memberPwRe">비밀번호 확인</label>
-                </div>
-                <div className="input-item">
-                  <input
-                    type="password"
-                    name="memberPwRe"
-                    id="memberPwRe"
-                    value={memberPwRe}
-                    onChange={inputMemberPwRe}
-                    onBlur={checkPw} // 입력이 끝나면 비밀번호 일치 여부 확인
-                    placeholder="위 비밀번호와 동일하게 입력"
-                  />
-                </div>
-                <p ref={pwMsgRef}></p>
-              </div>
-
-              {/* 닉네임 입력 */}
-              <div className="input-wrap">
-                <div className="input-title">
-                  <label htmlFor="memberNickname">닉네임</label>
-                </div>
-                <div className="input-item">
-                  <input
-                    type="text"
-                    name="memberNickname"
-                    id="memberNickname"
-                    value={member.memberNickname}
-                    onChange={inputMemberData}
-                    onBlur={checkNickname} // 닉네임이 입력되면 유효성 및 중복 체크
-                    placeholder="자신을 표현할 수 있는 이름을 지어주세요."
-                  />
-                </div>
-                <p
-                  className={
-                    nicknameCheck === 0
-                      ? ""
-                      : nicknameCheck === 1
-                      ? "input-ok"
-                      : "input-error"
-                  }
-                >
-                  {nicknameCheck === 0
-                    ? ""
-                    : nicknameCheck === 1
-                    ? "사용가능한 닉네임 입니다."
-                    : nicknameCheck === 2
-                    ? "닉네임은 한글, 영어 대/소문자, 숫자 입니다."
-                    : "이미 사용중인 닉네임 입니다."}
-                </p>
-              </div>
-
-              {/* 휴대폰 번호 입력 */}
-              <div className="input-wrap">
-                <div className="input-title">
-                  <label htmlFor="memberPhone">휴대폰 번호</label>
-                </div>
-                <div className="input-item">
-                  <input
-                    type="text"
-                    name="memberPhone"
-                    id="memberPhone"
-                    value={member.memberPhone}
-                    onChange={inputMemberData}
-                    onBlur={checkPhone} // 휴대폰 번호 입력 시 유효성 체크
-                    maxLength={13} // 010-0000-0000 형식에 맞게 제한
-                    placeholder="010-0000-0000"
-                  />
-                </div>
-                <p
-                  className={
-                    phoneError === 0
-                      ? ""
-                      : phoneError === 1
-                      ? "input-ok"
-                      : "input-error"
-                  }
-                >
-                  {phoneError === 0
-                    ? ""
-                    : phoneError === 1
-                    ? ""
-                    : "휴대폰 번호는 010-0000-0000 형식으로 입력해주세요."}
-                </p>
-              </div>
-
-              {/* 생년월일 입력 */}
-              <div className="input-wrap">
-                <div className="input-title">
-                  <label htmlFor="memberBirth">생년월일</label>
-                </div>
-                <div className="select-wrap">
-                  <div className="year">
-                    <select
-                      name="year"
-                      id="year"
-                      value={year}
-                      onChange={(e) => setYear(e.target.value)}
-                    >
-                      <option value="">년도</option>
-                      {years.map((yearOption) => (
-                        <option key={yearOption} value={yearOption}>
-                          {yearOption}
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                  <div className="month">
-                    <select
-                      name="month"
-                      id="month"
-                      value={month}
-                      onChange={(e) => setMonth(e.target.value)}
-                    >
-                      <option value="">월</option>
-                      {months.map((monthOption) => (
-                        <option key={monthOption} value={monthOption}>
-                          {monthOption}월
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                  <div className="day">
-                    <select
-                      name="day"
-                      id="day"
-                      value={day}
-                      onChange={(e) => setDay(e.target.value)}
-                    >
-                      <option value="">일</option>
-                      {days.map((dayOption) => (
-                        <option key={dayOption} value={dayOption}>
-                          {dayOption}일
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                </div>
-              </div>
-
-              {/* 성별 선택 */}
-              <div className="input-wrap">
-                <div className="input-title">
-                  <label htmlFor="memberGender">성별</label>
-                </div>
-                <div>
-                  <div>
-                    <label className="memberGender">
-                      <input
-                        type="radio"
-                        value="남"
-                        name="memberGender"
-                        id="memberGender"
-                        checked={member.memberGender === "남"}
-                        onChange={inputMemberData}
-                      />
-                      남성
-                    </label>
-                    <label className="memberGender">
-                      <input
-                        type="radio"
-                        value="여"
-                        name="memberGender"
-                        id="memberGender"
-                        checked={member.memberGender === "여"}
-                        onChange={inputMemberData}
-                      />
-                      여성
-                    </label>
-                  </div>
-                </div>
-              </div>
-
-              {/* 회원가입 버튼 */}
-              <div className="join-button-box2">
-                <button
-                  type="submit"
-                  className={`join-button ${isFormValid ? "" : "disabled"}`}
-                  disabled={!isFormValid}
-                  style={{
-                    pointerEvents: !isFormValid ? "none" : "auto", // 버튼 비활성화 시 클릭 불가
-                    backgroundColor: isFormValid ? "#30c272" : "white", // 비활성화 시 배경색 흰색으로
-                    color: isFormValid ? "white" : "#d3d3d3", // 비활성화 시 글자색 여린 회색으로
-                  }}
-                  onMouseEnter={(e) => {
-                    if (isFormValid) {
-                      e.target.style.backgroundColor = "#166139"; // hover 시 배경색 변경
-                    }
-                  }}
-                  onMouseLeave={(e) => {
-                    if (isFormValid) {
-                      e.target.style.backgroundColor = "#30c272"; // hover 끝난 후 원래 배경색으로 복구
-                    }
-                  }}
-                >
-                  회원가입
-                </button>
-              </div>
-            </form>
+              다음
+            </button>
           </div>
-        </>
-      ) : (
+        </div>
+      </>
+      {/*      ) : (
         <div className="join-title">
           <h3>코드 누락</h3>
           <p>코드가 인증되지 않았습니다.</p>
         </div>
-      )}
+      )} */}
     </section>
   );
 };
 
-export default Join2;
+export default Join3;
